+++ plotter/rings/rings.go
@@ -40,9 +40,8 @@ type LineStyleFunc func(interface{}) draw.LineStyle
 type Feature interface {
 	// Start and End indicate the position of the feature within the
 	// containing Location's coordinate system.
-	// TODO(kortschak): Make these return float64.
-	Start() int
-	End() int
+	Start() float64
+	End() float64
 
 	// Name returns the name of the feature.
 	Name() string
@@ -51,8 +50,7 @@ type Feature interface {
 	Location() Feature
 }
 
-// TODO(kortschak): Make this return float64.
-func lengthOf(f Feature) int {
+func lengthOf(f Feature) float64 {
 	return f.End() - f.Start()
 }
